requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
Flask-Cors==3.0.10
